refactored code using OOP refactored composer with psr-0 autoloadin, added bin script created a 'init_autoloader.php' to not bind classes with composer classloading added setter in Daux.php to let using daux.io as library
