+++ index.php
@@ -1,5 +1,5 @@
 <?php
-require_once("libs/daux.php");
+require_once("init_autoloader.php");
 
 /*
