+++ libs/Todaymade/Daux/Page.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Todaymade\Daux;
+
+
+interface Page {
+
+}
